backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
flask-cors==6.0.2
PyMySQL==1.1.2
pandas==3.0.1
openpyxl==3.1.5
Werkzeug==3.1.3
PyJWT==2.12.1
python-dotenv==1.2.1
